Struct sm4_firmware::staging::Serial

pub struct Serial<USART, PINS, WORD = u8> {
    usart: USART,
    pins: PINS,
    _word: PhantomData<WORD>,
}

Serial abstraction

Fields
usart: USARTpins: PINS_word: PhantomData<WORD>
Implementations

impl<USART, PINS, WORD> Serial<USART, PINS, WORD> where
PINS: Pins<USART>,
USART: Instance,

pub fn new(
usart: USART,
pins: PINS,
config: Config,
clocks: Clocks
) -> Result<Self, InvalidConfig>

pub fn listen(&mut self, event: Event)

Starts listening for an interrupt event

Note, you will also have to enable the corresponding interrupt in the NVIC to start receiving events.

pub fn unlisten(&mut self, event: Event)

Stop listening for an interrupt event

pub fn is_idle(&self) -> bool

Return true if the line idle status is set

pub fn is_txe(&self) -> bool

Return true if the tx register is empty (and can accept data)

pub fn is_rxne(&self) -> bool

Return true if the rx register is not empty (and can be read)

pub fn split(self) -> (Tx<USART, WORD>, Rx<USART, WORD>)

pub fn release(self) -> (USART, PINS)

impl<USART, PINS> Serial<USART, PINS, u8> where
PINS: Pins<USART>,
USART: Instance,

pub fn with_u16_data(self) -> Serial<USART, PINS, u16>

Converts this Serial into a version that can read and write u16 values instead of u8s

This can be used with a word length of 9 bits.

impl<USART, PINS> Serial<USART, PINS, u16> where
PINS: Pins<USART>,
USART: Instance,

pub fn with_u8_data(self) -> Serial<USART, PINS, u8>

Converts this Serial into a version that can read and write u8 values instead of u16s

This can be used with a word length of 8 bits.

impl<USART, PINS, WORD> Serial<USART, PINS, WORD> where
PINS: Pins<USART>,
USART: Instance,

fn config_stop(self, config: Config) -> Self

impl<USART, PINS> Serial<USART, PINS> where
PINS: Pins<USART>,
USART: Instance,

pub fn usart1(
usart: USART,
pins: PINS,
config: Config,
clocks: Clocks
) -> Result<Self, InvalidConfig>